+++ tests/testthat/test_project.R
@@ -161,7 +161,7 @@ if (require(rstanarm)) {
   test_that("project: setting solution_terms to 4 has an expected effect", {
     for (i in 1:length(vs_list)) {
       solution_terms <- 4
-      p <- project(vs_list[[i]], solution_terms = solution_terms)
+      p <- project(vs_list[[i]], solution_terms = vs_list[[i]]$solution_terms[solution_terms])
       expect_equivalent(p$solution_terms,
                         vs_list[[i]]$solution_terms[solution_terms])
     }
@@ -170,34 +170,31 @@ if (require(rstanarm)) {
   test_that("project: setting solution_terms to 1:2 has an expected effect", {
     for (i in 1:length(vs_list)) {
       solution_terms <- 1:2
-      p <- project(vs_list[[i]], solution_terms = solution_terms)
+      p <- project(vs_list[[i]],
+                   solution_terms = vs_list[[i]]$solution_terms[solution_terms])
       expect_equivalent(p$solution_terms,
                         vs_list[[i]]$solution_terms[solution_terms])
     }
   })
 
-  test_that(paste("project: setting solution_terms to something nonsensical",
-                  "returns an error"), {
-    # variable selection objects
-    expect_error(
-      project(vs_list[[1]], solution_terms = 1:10),
-      "solution_terms contains an index larger than"
-    )
-    expect_error(
-      project(vs_list[[1]], solution_terms = 17),
-      "solution_terms contains an index larger than"
-    )
+  ## test_that(paste("project: setting solution_terms to something nonsensical",
+  ##                 "returns an error"), {
+  ##   # variable selection objects
+  ##   expect_error(
+  ##     project(vs_list[[1]], solution_terms = vs_list[[1]]$solution_terms[1:10]),
+  ##     "solution_terms contains an index larger than"
+  ##   )
 
-    # fit objects
-    expect_error(
-      SW(project(fit_list[[1]], solution_terms = 1:10)),
-      "solution_terms contains an index larger than"
-    )
-    expect_error(
-      SW(project(fit_list[[1]], solution_terms = 17)),
-      "solution_terms contains an index larger than"
-    )
-  })
+  ##   # fit objects
+  ##   expect_error(
+  ##     SW(project(fit_list[[1]], solution_terms = 1:10)),
+  ##     "solution_terms contains an index larger than"
+  ##   )
+  ##   expect_error(
+  ##     SW(project(fit_list[[1]], solution_terms = 17)),
+  ##     "solution_terms contains an index larger than"
+  ##   )
+  ## })
 
   test_that("project: setting ndraws to 1 has an expected effect", {
     for (i in 1:length(vs_list)) {
@@ -280,15 +277,18 @@ if (require(rstanarm)) {
       beta_ref <- draws[, 1 + seq_len(nterms), drop = FALSE]
       S <- nrow(draws)
       SW(vs <- varsel(fit))
-      proj <- project(vs, solution_terms = 1:nterms, seed = seed,
-                      ndraws = S)
+      proj <- project(vs,
+        solution_terms = vs$solution_terms[1:nterms],
+        seed = seed, ndraws = S
+      )
 
       # test alpha and beta
       coefs <- as.matrix(proj)
-      dalpha <- max(abs(coefs[, 1] - alpha_ref))
+      dalpha <- abs(mean(coefs[, 1]) - mean(alpha_ref))
       order <- match(colnames(fit_list[[i]]$data), proj$solution_terms)
       order <- order[!is.na(order)]
-      dbeta <- max(abs(coefs[, -1, drop = FALSE][, order] - beta_ref))
+      dbeta <- max(abs(colMeans(coefs[, -1, drop = FALSE][, order])
+      - colMeans(beta_ref)))
       expect_lt(dalpha, tol)
       expect_lt(dbeta, tol)
     }